requirements.txt
google
google.genai
os
requests
bs4
dotenv
selenium.webdriver.chrome.options
selenium
time
flask
base64
flask_cors
pillow
io